fine_tune_llm/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-finetuning"
version = "0.1.0"
description = "LLM fine-tuning with LoRA for bird flu classification"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Core ML libraries for LLM fine-tuning
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "peft>=0.11.0",
    "trl>=0.7.0",
    "datasets>=2.14.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.43.0",
    # Evaluation and monitoring
    "evaluate>=0.4.0",
    "tensorboard>=2.14.0",
    "wandb>=0.15.0",
    # Hyperparameter tuning
    "optuna>=3.3.0",
    # UI
    "gradio>=4.0.0",
    # Data augmentation
    "nlpaug>=1.1.11",
    # Structured output
    "outlines>=0.0.30",
    # Data handling
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.19.0",
    "scikit-learn>=1.3.0",
    # Development tools
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
strict_optional = true